Move hooks to dedicated directory
